nlp-service/requirements.txt
fastapi
uvicorn
torch
transformers
pydantic
python-dotenv
scikit-learn
